Replace unsigned int with size_t for FastSourceLineResover load
as https://chromium-review.googlesource.com/c/breakpad/breakpad/+/4477821
demostrates, using size_t instead for symbols large than 4GB, otherwise it'll failed.

+++ src/processor/fast_source_line_resolver.cc
@@ -230,18 +230,18 @@ bool FastSourceLineResolver::Module::LoadMapFromMemory(
 
   const uint32_t* map_sizes = reinterpret_cast<const uint32_t*>(mem_buffer);
 
-  unsigned int header_size = kNumberMaps_ * sizeof(unsigned int);
+  size_t header_size = kNumberMaps_ * sizeof(unsigned int);
 
   // offsets[]: an array of offset addresses (with respect to mem_buffer),
   // for each "Static***Map" component of Module.
   // "Static***Map": static version of std::map or map wrapper, i.e., StaticMap,
   // StaticAddressMap, StaticContainedRangeMap, and StaticRangeMap.
-  unsigned int offsets[kNumberMaps_];
+  size_t offsets[kNumberMaps_];
   offsets[0] = header_size;
   for (int i = 1; i < kNumberMaps_; ++i) {
     offsets[i] = offsets[i - 1] + map_sizes[i - 1];
   }
-  unsigned int expected_size = sizeof(bool) + offsets[kNumberMaps_ - 1] +
+  size_t expected_size = sizeof(bool) + offsets[kNumberMaps_ - 1] +
                                map_sizes[kNumberMaps_ - 1] + 1;
   if (expected_size != memory_buffer_size &&
       // Allow for having an extra null terminator.
